+++ lib/src/default_progress_loader_widget.dart
@@ -40,7 +40,17 @@ class _DefaultProgressLoaderWidgetState
     super.dispose();
   }
 
-  Future<void> dismiss() => animationController.reverse();
+  /// Important note:
+  /// [AnimationController].reverse() does not return a [Future], it returns a [TickerFuture].
+  /// And [TickerFuture] will not return (and will hang forever) in some cases,
+  /// like if the [AnimationController] is dismissed before the [TickerFuture] completes for example.
+  /// To avoid hanging the [ProgressLoader] we must use [TickerFuture].orCancel and catch any [TickerCanceled] exception.
+  Future<void> dismiss() => animationController.reverse().orCancel.catchError(
+        (Object _) {
+          /// The error is a [TickerCanceled], just complete the future.
+        },
+        test: (error) => error is TickerCanceled,
+      );
 
   @override
   Widget build(BuildContext context) => AnimatedBuilder(